Refactor teachers to vit_package (#79)
* Refactor teachers to vit_package

* Fix format

* Simplify loading logic
