+++ src/lightly_train/_models/dinov2_vit/dinov2_vit_src/configs/__init__.py
@@ -15,35 +15,36 @@ from omegaconf import OmegaConf
 
 MODELS = {
     "vits14": {
-        "url": "https://dl.fbaipublicfiles.com/dinov2/dinov2_vits14/dinov2_vits14_pretrain.pth",
-        "config": "eval/vits14_pretrain",
-    },
-    "vitb14": {
-        "url": "https://dl.fbaipublicfiles.com/dinov2/dinov2_vitb14/dinov2_vitb14_pretrain.pth",
-        "config": "eval/vitb14_pretrain",
-    },
-    "vitl14": {
-        "url": "https://dl.fbaipublicfiles.com/dinov2/dinov2_vitl14/dinov2_vitl14_pretrain.pth",
-        "config": "eval/vitl14_pretrain",
-    },
-    "vitg14": {
-        "url": "https://dl.fbaipublicfiles.com/dinov2/dinov2_vitg14/dinov2_vitg14_pretrain.pth",
-        "config": "eval/vitg14_pretrain",
-    },
-}
-
-TRAIN_MODELS = {
-    "vits14": {
+        "url": "",
         "config": "train/vits14",
     },
     "vitb14": {
+        "url": "",
         "config": "train/vitb14",
     },
     "vitl14": {
+        "url": "",
         "config": "train/vitl14",
     },
     "vitg14": {
+        "url": "",
         "config": "train/vitg14",
+    },
+    "vits14_pretrain": {
+        "url": "https://dl.fbaipublicfiles.com/dinov2/dinov2_vits14/dinov2_vits14_pretrain.pth",
+        "config": "eval/vits14_pretrain",
+    },
+    "vitb14_pretrain": {
+        "url": "https://dl.fbaipublicfiles.com/dinov2/dinov2_vitb14/dinov2_vitb14_pretrain.pth",
+        "config": "eval/vitb14_pretrain",
+    },
+    "vitl14_pretrain": {
+        "url": "https://dl.fbaipublicfiles.com/dinov2/dinov2_vitl14/dinov2_vitl14_pretrain.pth",
+        "config": "eval/vitl14_pretrain",
+    },
+    "vitg14_pretrain": {
+        "url": "https://dl.fbaipublicfiles.com/dinov2/dinov2_vitg14/dinov2_vitg14_pretrain.pth",
+        "config": "eval/vitg14_pretrain",
     },
 }
