+++ src/lightly_train/_models/dinov2_vit/dinov2_vit_src/configs/eval/vitl14_reg4_pretrain.yaml
