+++ src/lightly_train/_models/dinov2_vit/dinov2_vit_src/configs/eval/vits14_reg4_pretrain.yaml
