+++ src/lightly_train/_models/dinov2_vit/dinov2_vit_src/dinov2_helper.py
@@ -13,7 +13,7 @@ import torch
 
 from lightly_train import _distributed as distributed_helpers
 from lightly_train._data.download import download_from_url
-from lightly_train._modules.teachers.dinov2.models.vision_transformer import (
+from lightly_train._models.dinov2_vit.dinov2_vit_src.models.vision_transformer import (
     DinoVisionTransformer,
 )
