+++ src/lightly_train/_models/dinov2_vit/dinov2_vit_src/layers/__init__.py
@@ -1,0 +1,30 @@
+#
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+#
+# This source code is licensed under the Apache License, Version 2.0
+# found in the LICENSE file in the root directory of this source tree.
+#
+
+from lightly_train._models.dinov2_vit.dinov2_vit_src.layers.attention import (
+    MemEffAttention,
+)
+from lightly_train._models.dinov2_vit.dinov2_vit_src.layers.block import (
+    NestedTensorBlock,
+)
+from lightly_train._models.dinov2_vit.dinov2_vit_src.layers.mlp import Mlp
+from lightly_train._models.dinov2_vit.dinov2_vit_src.layers.patch_embed import (
+    PatchEmbed,
+)
+from lightly_train._models.dinov2_vit.dinov2_vit_src.layers.swiglu_ffn import (
+    SwiGLUFFN,
+    SwiGLUFFNFused,
+)
+
+__all__ = [
+    "MemEffAttention",
+    "NestedTensorBlock",
+    "Mlp",
+    "PatchEmbed",
+    "SwiGLUFFN",
+    "SwiGLUFFNFused",
+]
