+++ src/lightly_train/_models/dinov2_vit/dinov2_vit_src/layers/layer_scale.py
