--- src/lightly_train/_modules/teachers/dinov2/__init__.py
@@ -1,6 +1,0 @@
-#
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-#
-# This source code is licensed under the Apache License, Version 2.0
-# found in the LICENSE file in the root directory of this source tree.
-#
