--- src/lightly_train/_modules/teachers/dinov2/layers/__init__.py
@@ -1,24 +1,0 @@
-#
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-#
-# This source code is licensed under the Apache License, Version 2.0
-# found in the LICENSE file in the root directory of this source tree.
-#
-
-from lightly_train._modules.teachers.dinov2.layers.attention import MemEffAttention
-from lightly_train._modules.teachers.dinov2.layers.block import NestedTensorBlock
-from lightly_train._modules.teachers.dinov2.layers.mlp import Mlp
-from lightly_train._modules.teachers.dinov2.layers.patch_embed import PatchEmbed
-from lightly_train._modules.teachers.dinov2.layers.swiglu_ffn import (
-    SwiGLUFFN,
-    SwiGLUFFNFused,
-)
-
-__all__ = [
-    "MemEffAttention",
-    "NestedTensorBlock",
-    "Mlp",
-    "PatchEmbed",
-    "SwiGLUFFN",
-    "SwiGLUFFNFused",
-]
