+++ tests/_models/dinov2_vit/test_dinov2_vit.py
@@ -12,8 +12,8 @@ import torch
 
 from lightly_train._models.dinov2_vit.dinov2_vit import DINOv2ViTModelWrapper
 from lightly_train._models.dinov2_vit.dinov2_vit_package import DINOv2ViTPackage
-from lightly_train._modules.teachers.dinov2.layers.drop_path import DropPath
-from lightly_train._modules.teachers.dinov2.models.vision_transformer import (
+from lightly_train._models.dinov2_vit.dinov2_vit_src.layers.drop_path import DropPath
+from lightly_train._models.dinov2_vit.dinov2_vit_src.models.vision_transformer import (
     vit_small as vit_small,
 )
